+++ packages/common/lib/models/ReactionMessage.ts
@@ -1,0 +1,10 @@
+import { Entity, PrimaryColumn, Column } from 'typeorm';
+
+@Entity('reaction_messages')
+export class ReactionMessage {
+  @PrimaryColumn('bigint')
+  public id!: string;
+
+  @Column()
+  public unique!: boolean;
+}